Call toDouble in DoubleProperty
This prevent to hit an assertion when decoding a double encoded as an integer in the JSON.

+++ lib/src/config/property.dart
@@ -152,7 +152,7 @@ class DoubleProperty extends ApiConfigSchemaProperty {
   _singleRequestValue(value) {
     assert(value != null);
     if (value is num) {
-      return value;
+      return value.toDouble();
     }
     try {
       return double.parse(value);
